Refactor getTasks API into TaskService class
Thinking about the PUT method drove me to think about persistance
so I decided to change the getTasks function I've created and include
it as part of a TaskService class.

Doing so allows me to create a TaskService instance in the express
routine defined in start.js

This new object can "hold" the information of the tasks that have been
requested while the application is running thus, providing a rudimentary
form of idempotence.